docs: add .prettierrc configuration
It is a default configuration for Prettier - opinionated code formatter,
that can be used to format .md and .mdx files.

The only non-default config is HTML whitespace sensitivity.
The default value is to respect CSS setting, but we will use it for
other purposes, mostly for markdown and MDX, and thus we don't need to
preserve spaces that much